requirements.txt
matplotlib==3.5.2
numpy==1.22.3
Pillow==9.4.0
scikit_learn==1.2.2
scipy==1.8.0
seaborn==0.11.2
six==1.14.0
tensorboard_logger==0.1.0
torch==1.13.0
torchvision==0.14.0
xlrd==1.2.0
xlwt==1.3.0
tensorflow==2.11.0
faiss-gpu==1.7.4
